Crate lbucket

Structs

ID

node unique identifier, useful when running in cluster mode

PREFIX_BUCKET

bucket key prefix

PREFIX_COUNTER

counter/captcha key prefix

Constants

BUCKET_EXPIRY_OFFSET

If buckets perform clean up at x instant, then buckets themselves will get cleaned up at x + BUCKET_EXPIRY_OFFSET(if they haven’t already been cleaned up)

HIT_PER_SECOND

Initial allocation ammount of buckets[bucket::Bucket]

PKG_NAME

PKG_VERSION

PREFIX_BUCKET_TIMER

bucket timer key prefix

REDIS_LBUCKET_BUCKET_TYPE_VERSION

Bucket[bucket::Bucket] type version

Functions

RedisModule_OnLoad

RedisModule_OnUnload